Build flannel and bridge plugin
Flannel plugin is designed to work in conjunction with
flannel, a network fabric for containers. When flannel
daemon is started, it outputs a /run/flannel/subnet.env
file that looks like this:

FLANNEL_NETWORK=10.1.0.0/16
FLANNEL_SUBNET=10.1.17.1/24
FLANNEL_MTU=1472
FLANNEL_IPMASQ=true

This information reflects the attributes of flannel
network on the host. The flannel CNI plugin uses this
information to configure another CNI plugin, such as
bridge plugin. And it delegates to invoke bridge plugin
to configure NICs for containers.

An example of flannel.conf is as follows

{
    "name": "mynet",
    "type": "galaxy-flannel",
    "delegate": {
        "type": "galaxy-bridge",
        "mtu": 1400
    },
    "subnetFile": "/run/flannel/subnet.env"
}